cdrom: mount the data track from multi-FILE cues (fixes SystemHalt(906))
The cue parser let every FILE line overwrite the mounted image, so a
"one .bin per track" rip (data track + CD-DA tracks — the common rip
layout) mounted the LAST audio track as the data disc. The BIOS then
could not Load() the boot EXE and panicked:

    if ( Load (0xA000B8B0, 0xA000B870 ) == 0 )
        SystemHalt (906);          // docs/psx_bios_disasm.txt:600

Confirmed on Brave Fencer Musashi (SLUS-00726), whose disc is a 4-FILE
cue: boot reached the Sony logo, then rendered a corrupted second logo
(audio bytes decoded as texture data) and halted. Registers at the halt
match the disasm exactly — a0=0x42 ('B'), i.e. SystemError('B', 906).

Fix, general to any multi-FILE cue:
  - The mounted image is the FIRST file (the data track), not the last.
  - Absolute track LBAs accumulate the sector counts of preceding files,
    since each INDEX time in a multi-FILE cue is relative to its own file.
Single-FILE cues are unchanged: one file, INDEX already absolute, base 0.

With this, BFM boots through both logos to the title screen from the
.cue with no CLI override.

Also here:
  - debug_server: pass uint64_t (not unsigned long long) to
    overlay_sljit_get_status — the mismatch is a hard error on LP64 gcc
    and broke the Linux build outright.
  - runtime: register the musashi-runtime target, guarded on the
    generated C existing so a clean clone still configures.
  - tools: mark bios_emitter_fingerprint.sh executable (regen_bios.sh
    invokes it directly and failed on a fresh checkout).

Regression guard: runtime/tests/test_cue_multifile.cpp — asserts the
data track is mounted and the LBA accumulation is right; fails on the
pre-fix parser.

+++ runtime/src/debug_server.c
@@ -9930,7 +9930,7 @@ static void handle_sljit_status(int id, const char *json)
     (void)json;
     int selftest = overlay_sljit_selftest();
     int available = 0, st_ok = 0;
-    unsigned long long compiles = 0, declines = 0, bytes = 0;
+    uint64_t compiles = 0, declines = 0, bytes = 0;
     overlay_sljit_get_status(&available, &st_ok, &compiles, &declines, &bytes);
     send_fmt("{\"id\":%d,\"ok\":true,\"backend\":\"%s\",\"available\":%d,"
              "\"selftest_ok\":%d,\"live\":%d,\"compiles\":%llu,\"declines\":%llu,"
@@ -9939,7 +9939,8 @@ static void handle_sljit_status(int id, const char *json)
              "\"persist_on\":%d,\"persist_calls\":%u,\"persist_writes\":%u,"
              "\"persist_dbg\":%d}\n",
              id, overlay_backend_name(overlay_backend_active()), available,
-             selftest, overlay_loader_get_sljit_live(), compiles, declines, bytes,
+             selftest, overlay_loader_get_sljit_live(),
+             (unsigned long long)compiles, (unsigned long long)declines, (unsigned long long)bytes,
              overlay_loader_sljit_registered(),
              overlay_loader_sljit_obsoleted(),
              overlay_loader_sljit_reloaded(),
